refactor: extract bp_usage to routes/usage.py
Phase 5.3 of the dashboard.py modularisation. Moves the 12 usage / anomaly
/ attribution routes registered on bp_usage into routes/usage.py, following
the pattern established by routes/sessions.py and routes/brain.py.

Module-level helpers (_usage_cache, _compute_transcript_analytics,
_detect_and_store_anomalies, _get_anomaly_db, _build_cost_comparison, etc.)
stay in dashboard.py and are reached via late `import dashboard as _d`.
Pure mechanical move — zero behaviour change.

Net: dashboard.py 32,746 -> 31,729 lines; new routes/usage.py = 1,069 lines.
All 12 endpoints return 200 under smoke test.
